// codebase/interfaces.ts
export interface Options {
    isTextExact?: boolean;
    asDisplayed?: boolean;
    isReplace?: boolean;
    isPaste?: boolean;
    isNumber?: boolean;
    offsetX?: number;
    offsetY?: number;
    timeout?: number;
    interval?: number;
    waitSpinner?: number;
    readOnlyMode?: boolean;
    wait?: boolean;
    pattern?: string;
    reduceOffset?: boolean;
    isSkipped?: boolean;
    closeConnection?: boolean;
    checkErrors?: boolean;
    cache?: boolean;
    skipLogging?: boolean;
    retry?: boolean;
}